requirements.txt
Brotli==1.0.9
click==8.0.3
colorama==0.4.4
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
pandas==1.3.5
plotly==5.5.0
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
six==1.16.0
tenacity==8.0.1
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.7.0
dash_charts-0.0.1.tar.gz